up to 4.22

+++ osa4/blogilista/controllers/blogsRouter.js
@@ -1,16 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
-const config = require('../utils/config')
-
-const getTokenFrom = (request) => {
-  const authHeader = request.get('authorization')
-  if (authHeader && authHeader.toLowerCase().startsWith('bearer ')) {
-    return authHeader.substring(7)
-  }
-  return null
-}
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', (request, response) => {
   Blog
@@ -21,19 +12,13 @@ blogsRouter.get('/', (request, response) => {
     })
 })
 
-blogsRouter.post('/', (request, response, next) => {
+blogsRouter.post('/', userExtractor, (request, response, next) => {
   let tempBlog = request.body
   if ( !('likes' in tempBlog) ) {
     tempBlog = { ...tempBlog, likes: 0 }
   }
 
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, config.SECRET)
-  if (!token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
-
-  User.findById(decodedToken.id)
+  User.findById(request.user)
     .then( result => {
       tempBlog = { ...tempBlog, user: result._id }
       const blog = new Blog(tempBlog)
@@ -47,16 +32,21 @@ blogsRouter.post('/', (request, response, next) => {
     .catch(error => next(error))
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
   try {
-    const result = await Blog.findByIdAndDelete(request.params.id)
-    response.status(200).json(result)
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === request.user) {
+      const result = await Blog.findByIdAndDelete(request.params.id)
+      response.status(200).json(result)
+    } else {
+      response.status(401).json({ error: 'You are not the owner of this blog entry' })
+    }
   } catch (exception) {
     next(exception)
   }
 })
 
-blogsRouter.put('/:id', async (request, response, next) => {
+blogsRouter.put('/:id', userExtractor, async (request, response, next) => {
   try {
     const modifiedBlog = request.body
     const result = await Blog.findByIdAndUpdate(request.params.id, modifiedBlog, { new: true })
